# sdk/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Open Configuration Management API",
    "version": "v1",
    "description": "Multi-tenant configuration management with versioning and access control"
  },
  "paths": {
...
    "/api/v1/~config-metadata-schema": {
      "get": {
        "operationId": "get_config_metadata_schema",
        "summary": "Config metadata block JSON Schema",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response",
                  "type": "object"
                }
              }
            }
          }
        },
        "description": "Return JSON Schema for the top-level config metadata block (e.g. ``_ocmo``).",
        "tags": [
          "Config"
        ],
        "security": [
          {
            "OAuth2AuthorizationCodeBearer": []
          }
        ]
      }
    },
    "/api/v1/ns/{namespace}/~config-schema/{path}": {
      "get": {
        "operationId": "get_config_data_schema",
        "summary": "Config data JSON Schema",
        "parameters": [
          {
            "in": "path",
            "name": "namespace",
            "schema": {
              "title": "Namespace",
              "type": "string"
            },
            "required": true
          },
          {
            "in": "path",
            "name": "path",
            "schema": {
              "title": "Path",
              "type": "string"
            },
            "required": true
          },
          {
            "in": "query",
            "name": "version",
            "schema": {
              "default": "latest",
              "title": "Version",
              "type": "string"
            },
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response",
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorSchema"
                }
              }
            }
          }
        },
        "description": "Return JSON Schema for a config's data body, if one is defined.",
        "tags": [
          "Config"
        ],
        "security": [
          {
            "OAuth2AuthorizationCodeBearer": []
          }
        ]
      }
    },
    "/api/v1/ns/{namespace}/~config/~create/{path}": {
      "post": {
        "operationId": "create_config",
        "summary": "Create Config",
        "parameters": [
          {
            "in": "path",
            "name": "namespace",
            "schema": {
              "title": "Namespace",
              "type": "string"
            },
            "required": true
          },
          {
            "in": "path",
            "name": "path",
            "schema": {
              "title": "Path",
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "discriminator": {
                    "mapping": {
                      "config": "#/components/schemas/ConfigSchemaExtended",
                      "folder": "#/components/schemas/FolderSchema",
                      "resolver": "#/components/schemas/ResolverSchema",
                      "secret": "#/components/schemas/SecretSchemaExtended",
                      "template": "#/components/schemas/TemplateSchemaExtended"
                    },
                    "propertyName": "node_type"
                  },
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ConfigSchemaExtended"
                    },
                    {
                      "$ref": "#/components/schemas/TemplateSchemaExtended"
                    },
                    {
                      "$ref": "#/components/schemas/SecretSchemaExtended"
                    },
                    {
                      "$ref": "#/components/schemas/ResolverSchema"
                    },
                    {
                      "$ref": "#/components/schemas/FolderSchema"
                    }
                  ],
                  "title": "Response"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorSchema"
                }
              }
            }
          },
          "413": {
            "description": "Content Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorSchema"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorSchema"
                }
              }
            }
          }
        },
        "description": "Create a new Config at the specified path.",
        "tags": [
          "Config"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigDocument",
                "description": "Config YAML document.",
                "format": "textarea",
                "title": "ConfigDocument",
                "type": "string"
              },
              "example": {
                "foo": "bar",
                "baz": 42
              }
            },
            "application/yaml": {
              "schema": {
                "description": "Config YAML document.",
                "format": "textarea",
                "title": "ConfigDocument",
                "type": "string"
              },
              "example": "foo: bar\nbaz: 42"
            },
            "application/octet-stream": {
              "title": "File upload",
              "schema": {
                "type": "string",
                "format": "binary"
              },
              "description": "Upload config as a file",
              "example": "config.yaml"
            }
          },
          "required": true,
          "description": "Document as raw body: pick a text Content-Type to paste, or application/octet-stream to upload a file"
        },
        "security": [
          {
            "OAuth2AuthorizationCodeBearer": []
          }
        ]
      }
    },
    "/api/v1/ns/{namespace}/~config/~update/{path}": {
      "put": {
        "operationId": "update_config",
        "summary": "Update Config",
        "parameters": [
          {
            "in": "path",
            "name": "namespace",
            "schema": {
              "title": "Namespace",
              "type": "string"
            },
            "required": true
          },
          {
            "in": "path",
            "name": "path",
            "schema": {
              "title": "Path",
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "discriminator": {
                    "mapping": {
                      "config": "#/components/schemas/ConfigSchemaExtended",
                      "folder": "#/components/schemas/FolderSchema",
                      "resolver": "#/components/schemas/ResolverSchema",
                      "secret": "#/components/schemas/SecretSchemaExtended",
                      "template": "#/components/schemas/TemplateSchemaExtended"
                    },
                    "propertyName": "node_type"
                  },
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ConfigSchemaExtended"
                    },
                    {
                      "$ref": "#/components/schemas/TemplateSchemaExtended"
                    },
                    {
                      "$ref": "#/components/schemas/SecretSchemaExtended"
                    },
                    {
                      "$ref": "#/components/schemas/ResolverSchema"
                    },
                    {
                      "$ref": "#/components/schemas/FolderSchema"
                    }
                  ],
                  "title": "Response"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorSchema"
                }
              }
            }
          },
          "413": {
            "description": "Content Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorSchema"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorSchema"
                }
              }
            }
          }
        },
        "description": "Update an existing Config. Creates a new version only when content changes.",
        "tags": [
          "Config"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigDocument",
                "description": "Config YAML document.",
                "format": "textarea",
                "title": "ConfigDocument",
                "type": "string"
              },
              "example": {
                "foo": "bar",
                "baz": 42
              }
            },
            "application/yaml": {
              "schema": {
                "description": "Config YAML document.",
                "format": "textarea",
                "title": "ConfigDocument",
                "type": "string"
              },
              "example": "foo: bar\nbaz: 42"
            },
            "application/octet-stream": {
              "title": "File upload",
              "schema": {
                "type": "string",
                "format": "binary"
              },
              "description": "Upload config as a file",
              "example": "config.yaml"
            }
          },
          "required": true,
          "description": "Document as raw body: pick a text Content-Type to paste, or application/octet-stream to upload a file"
        },
        "security": [
          {
            "OAuth2AuthorizationCodeBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorSchema": {
        "properties": {
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "items": {},
                "type": "array"
              }
            ],
            "title": "Error"
          },
          "audit_event_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Audit log event ID when the failure was recorded server-side",
            "title": "Audit Event Id"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorSchema",
        "type": "object"
      },
...
      "ConfigSchemaExtended": {
        "properties": {
          "name": {
            "description": "Leaf segment name",
            "title": "Name",
            "type": "string"
          },
          "path": {
            "description": "Full path within the namespace",
            "title": "Path",
            "type": "string"
          },
          "node_type": {
            "const": "config",
            "title": "Node Type",
            "type": "string"
          },
          "author": {
            "description": "Last author identifier",
            "title": "Author",
            "type": "string"
          },
          "description": {
            "description": "Markdown description",
            "title": "Description",
            "type": "string"
          },
          "tags": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Tags",
            "type": "object"
          },
          "version_data": {
            "$ref": "#/components/schemas/VersionSchema"
          },
          "propagation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PropagationResultSchema"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "name",
          "path",
          "node_type",
          "author",
          "description",
          "tags",
          "version_data"
        ],
        "title": "ConfigSchemaExtended",
        "type": "object"
      },
      "FolderSchema": {
        "properties": {
          "name": {
            "description": "Leaf segment name",
            "title": "Name",
            "type": "string"
          },
          "path": {
            "description": "Full path within the namespace",
            "title": "Path",
            "type": "string"
          },
          "node_type": {
            "const": "folder",
            "title": "Node Type",
            "type": "string"
          },
          "author": {
            "description": "Last author identifier",
            "title": "Author",
            "type": "string"
          },
          "description": {
            "description": "Markdown description",
            "title": "Description",
            "type": "string"
          },
          "created_at": {
            "description": "Creation timestamp",
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          }
        },
        "required": [
          "name",
          "path",
          "node_type",
          "author",
          "description",
          "created_at"
        ],
        "title": "FolderSchema",
        "type": "object"
      },
      "PropagationResultSchema": {
        "properties": {
          "source_path": {
            "title": "Source Path",
            "type": "string"
          },
          "source_version": {
            "title": "Source Version",
            "type": "integer"
          },
          "trigger": {
            "title": "Trigger",
            "type": "string"
          },
          "targets": {
            "items": {
              "$ref": "#/components/schemas/PropagationTargetResult"
            },
            "title": "Targets",
            "type": "array"
          }
        },
        "required": [
          "source_path",
          "source_version",
          "trigger",
          "targets"
        ],
        "title": "PropagationResultSchema",
        "type": "object"
      },
      "PropagationTargetResult": {
        "properties": {
          "path": {
            "title": "Path",
            "type": "string"
          },
          "status": {
            "enum": [
              "updated",
              "unchanged",
              "skipped",
              "error"
            ],
            "title": "Status",
            "type": "string"
          },
          "version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "required": [
          "path",
          "status"
        ],
        "title": "PropagationTargetResult",
        "type": "object"
      },
      "ResolverSchema": {
        "properties": {
          "name": {
            "description": "Leaf segment name",
            "title": "Name",
            "type": "string"
          },
          "path": {
            "description": "Full path within the namespace",
            "title": "Path",
            "type": "string"
          },
          "node_type": {
            "const": "resolver",
            "title": "Node Type",
            "type": "string"
          },
          "author": {
            "description": "Last author identifier",
            "title": "Author",
            "type": "string"
          },
          "description": {
            "description": "Markdown description",
            "title": "Description",
            "type": "string"
          },
          "created_at": {
            "description": "Creation timestamp",
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "configuration": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "format": "textarea",
            "title": "Configuration"
          },
          "token1": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token1"
          },
          "token1_last_used": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token1 Last Used"
          },
          "token2": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token2"
          },
          "token2_last_used": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token2 Last Used"
          }
        },
        "required": [
          "name",
          "path",
          "node_type",
          "author",
          "description",
          "created_at"
        ],
        "title": "ResolverSchema",
        "type": "object"
      },
      "SecretSchemaExtended": {
        "properties": {
          "name": {
            "description": "Leaf segment name",
            "title": "Name",
            "type": "string"
          },
          "path": {
            "description": "Full path within the namespace",
            "title": "Path",
            "type": "string"
          },
          "node_type": {
            "const": "secret",
            "title": "Node Type",
            "type": "string"
          },
          "author": {
            "description": "Last author identifier",
            "title": "Author",
            "type": "string"
          },
          "description": {
            "description": "Markdown description",
            "title": "Description",
            "type": "string"
          },
          "tags": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Tags",
            "type": "object"
          },
          "version_data": {
            "$ref": "#/components/schemas/SecretVersionSchema"
          }
        },
        "required": [
          "name",
          "path",
          "node_type",
          "author",
          "description",
          "tags",
          "version_data"
        ],
        "title": "SecretSchemaExtended",
        "type": "object"
      },
      "SecretVersionSchema": {
        "properties": {
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              }
            ],
            "title": "Version"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "updater": {
            "title": "Updater",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "deleted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted At"
          },
          "data": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          }
        },
        "required": [
          "version",
          "tags",
          "updater",
          "updated_at",
          "deleted_at"
        ],
        "title": "SecretVersionSchema",
        "type": "object"
      },
      "TemplateSchemaExtended": {
        "properties": {
          "name": {
            "description": "Leaf segment name",
            "title": "Name",
            "type": "string"
          },
          "path": {
            "description": "Full path within the namespace",
            "title": "Path",
            "type": "string"
          },
          "node_type": {
            "const": "template",
            "title": "Node Type",
            "type": "string"
          },
          "author": {
            "description": "Last author identifier",
            "title": "Author",
            "type": "string"
          },
          "description": {
            "description": "Markdown description",
            "title": "Description",
            "type": "string"
          },
          "tags": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Tags",
            "type": "object"
          },
          "version_data": {
            "$ref": "#/components/schemas/VersionSchema"
          }
        },
        "required": [
          "name",
          "path",
          "node_type",
          "author",
          "description",
          "tags",
          "version_data"
        ],
        "title": "TemplateSchemaExtended",
        "type": "object"
      },
      "VersionSchema": {
        "properties": {
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              }
            ],
            "title": "Version"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "data": {
            "title": "Data",
            "type": "string"
          },
          "updater": {
            "title": "Updater",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "deleted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted At"
          }
        },
        "required": [
          "version",
          "tags",
          "data",
          "updater",
          "updated_at",
          "deleted_at"
        ],
        "title": "VersionSchema",
        "type": "object"
      },
...
      "ConfigDocument": {
        "description": "Config YAML document.",
        "format": "textarea",
        "title": "ConfigDocument",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "ResolverAuth": {
        "type": "apiKey",
        "in": "query",
        "name": "token"
      },
      "OAuth2AuthorizationCodeBearer": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "scopes": {
              "openid": "",
              "profile": "",
              "email": "",
              "groups": ""
            },
            "authorizationUrl": "http://localhost:8080/dex/auth",
            "tokenUrl": "http://localhost:8080/dex/token"
          }
        }
      }
    }
  },
  "servers": []
}